move db cli to support crate
