+++ axum-on-rails/src/lib.rs
@@ -3,5 +3,6 @@ use bb8_postgres::PostgresConnectionManager;
 use tokio_postgres::NoTls;
 
 pub mod test_helpers;
+pub mod cli;
 
 pub type ConnectionPool = Pool<PostgresConnectionManager<NoTls>>;
